requirements.txt
pandas>=2.0
numpy>=1.26
pyarrow>=15.0
matplotlib>=3.8
seaborn>=0.13
statsmodels>=0.14
lifelines>=0.27
requests>=2.31
python-dotenv>=1.0
